Delete "no unsafe" claim from readme
"No unsafe" is misleading because there is unsafe code unconditionally
compiled into the crate.

It looks like this sentence in the readme was written in 408f42a3 and
"no unsafe" was true at that point, but then 48 hours later some unsafe
code was added in 8904b5e7 and has been there ever since. It seems like
an oversight that the readme was not updated.