Raise MSRV to 1.56
The `banners` example already requires Rust 1.55 to build. Going one version further to 1.56
silences the (harmless) warning about the `rust-version` key in the Cargo.toml being
unsupported and unlocks Rust 2021.

The example could instead be fixed to build with Rust 1.51, but don't bother and just bump the MSRV
since 1.56 is already one and a half years old at this point.