Make most of the API const-friendly
Since owo-colors does most of its work via the type system, it makes sense for
it to be possible to construct the types at compile time. I've rewritten a few
methods to make them const friendly (no &mut refs...), but otherwise this
appears to work great.

One change I've made is to drop the `Sized` requirement on the types that are
only accessed via pointer indirection. This allows users to write types like
`FgColorDisplay<Green, str>` which feels great. I can't drop the `Sized`
requirement on `OwoColorize` due to API compat, though, sadly.

Whatever bits are left out are not possible to convert over to const currently.
But enough of the API is available (including dynamic styles!) that I think
most user needs will be met through this.

Closes #62.